+++ admin-console/mock-server/auth/auth.mock.js
@@ -1,0 +1,13 @@
+module.exports = [
+  {
+    request: {
+      method: 'POST',
+      path: '/login',
+    },
+    response: () => {
+      return {
+        statusCode: 200
+      }
+    }
+  }
+]