+++ admin-console/mock-server/index.js
@@ -1,12 +1,11 @@
-const project = require('../config/project.config')
-const jsonServer = require('json-server')
+const { runServer } = require('express-mock-server');
 
-const server = jsonServer.create()
-const router = jsonServer.router('./mock-server/db.json')
-const middlewares = jsonServer.defaults()
+const opt_serverConfig = {
+  port: 8080
+}
 
-server.use(middlewares)
-server.use(project.mock_server_prefix, router)
-server.listen(project.mock_server_port, function () {
-  console.log('Mock Server is running')
-})
+const sources = [
+  require('./auth/auth.mock')
+]
+
+runServer(sources, opt_serverConfig)