+++ admin-console/src/common/components/Item/item.styles.js
@@ -1,6 +1,6 @@
 import tc from 'tinycolor2'
-import { darkColor } from '../../styles/colors'
-import { itemSize, gap } from '../../styles/constants'
+import { darkColor } from '../../../styles/colors'
+import { itemSize, gap } from '../../../styles/constants'
 
 export default {
   main: {
@@ -9,11 +9,14 @@ export default {
     lineHeight: `${itemSize}px`,
     width: '100%',
     cursor: 'pointer',
-    paddingLeft: gap,
     boxSizing: 'border-box',
     color: tc(darkColor).lighten(50),
     backgroundColor: 'transparent',
     whiteSpace: 'nowrap'
+  },
+  label: {
+    paddingLeft: gap,
+    paddingRight: gap
   },
   arrow: {
     float: 'right',