+++ admin-console/src/routes/index.js
@@ -1,11 +1,15 @@
 import CoreLayout from '../layouts/CoreLayout/containers/coreLayout.container'
+import { initAuth } from '../common/actions/auth.actions'
 import Home from './Home'
 
 export const createRoutes = (store) => ({
-  path        : '/',
-  component   : CoreLayout,
-  indexRoute  : Home,
-  childRoutes : []
+  path: '/',
+  component: CoreLayout,
+  indexRoute: Home,
+  childRoutes: [],
+  onEnter: () => {
+    store.dispatch(initAuth())
+  }
 })
 
 export default createRoutes