+++ admin-console/src/store/reducers.js
@@ -1,11 +1,14 @@
 import { combineReducers } from 'redux'
 import locationReducer from './location'
+
 import coreLayoutReducer from '../layouts/CoreLayout/reducers/coreLayout.reducer'
+import authReducer from '../common/reducers/auth.reducer'
 
 export const makeRootReducer = (asyncReducers) => {
   return combineReducers({
     location: locationReducer,
     coreLayout: coreLayoutReducer,
+    auth: authReducer,
     ...asyncReducers
   })
 }