+++ admin-console/CHANGELOG.md
@@ -1,3 +1,10 @@
+**v0.0.3**
+----------
+- Added new mock server (express)
+- Added login mock
+- Added login modal and form
+
+
 **v0.0.2**
 ---------
 - Added Item Component